chore: flatten simulation module
